fix some oss-fuzz (#1184)
* fix oss-fuzz 10419.

* fix oss-fuzz 10427.

* fix oss-fuzz 10421.

* fix oss-fuzz 10422.

* fix oss-fuzz 10425.

* fix oss-fuzz 10426.

* fix oss-fuzz 10426.

* fix oss-fuzz 10422.

* fix oss-fuzz  10426.

* fix oss-fuzz 10456.

* fix oss-fuzz 10428.

* fix oss-fuzz 10429.

* fix oss-fuzz 10431.

* fix oss-fuzz 10435.

* fix oss-fuzz 10430.

* fix oss-fuzz 10436.

* remove unused var.

* fix oss-fuzz 10449.

* fix oss-fuzz 10452.

* fix oss-fuzz 11792.

* fix oss-fuzz 10457.

* fix oss-fuzz 11737.

* fix oss-fuzz 10458.

* fix oss-fuzz 10565.

* fix oss-fuzz 11651.

* fix oss-fuzz 10497.

* fix oss-fuzz 10515.

* fix oss-fuzz 10586.

* fix oss-fuzz 10597.

* fiz oss-fuzz 11721.

* fix oss-fuzz 10718.

* fix oss-fuzz 15610.

* fix oss-fuzz 10512.

* fix oss-fuzz 10545.

* fix oss-fuzz 10598.

* fix oss-fuzz 11112.

* fix oss-fuzz 11589.

* fix oss-fuzz 10674.

* git fix oss-fuzz 19610.

* fix oss-fuzz 19848.

* fix oss-fuzz 19851.

* fix oss-fuzz 19852.

* fix oss-fuzz 10878.

* fix oss-fuzz 11655.

* fix oss-fuzz 19849.

* fix oss-fuzz 11765.

* fix oss-fuzz 10337.

* fix oss-fuzz 10575.

* fix oss-fuzz 19877.

* fix oss-fuzz 19895.

* fix oss-fuzz 19896.

* fix oss-fuzz 19897.

* remove verbose fprintf output.

+++ qemu/target-m68k/helper.c
@@ -169,10 +169,10 @@ void HELPER(set_macsr)(CPUM68KState *env, uint32_t val)
             }
             if (env->macsr & MACSR_FI) {
                 regval = (((uint64_t)acc) << 8) | extlow;
-                regval |= ((int64_t)exthigh) << 40;
+                regval |= ((uint64_t)((int64_t)exthigh)) << 40;
             } else if (env->macsr & MACSR_SU) {
                 regval = acc | (((int64_t)extlow) << 32);
-                regval |= ((int64_t)exthigh) << 40;
+                regval |= ((uint64_t)((int64_t)exthigh)) << 40;
             } else {
                 regval = acc | (((uint64_t)extlow) << 32);
                 regval |= ((uint64_t)(uint8_t)exthigh) << 40;
@@ -609,7 +609,7 @@ void HELPER(macsatf)(CPUM68KState *env, uint32_t acc)
     int64_t result;
 
     sum = env->macc[acc];
-    result = (sum << 16) >> 16;
+    result = ((int64_t)((uint64_t)sum << 16)) >> 16;
     if (result != sum) {
         env->macsr |= MACSR_V;
     }
@@ -762,11 +762,11 @@ void HELPER(set_mac_exts)(CPUM68KState *env, uint32_t val, uint32_t acc)
     int32_t tmp;
     res = (uint32_t)env->macc[acc];
     tmp = (int16_t)val;
-    res |= ((int64_t)tmp) << 32;
+    res |= ((uint64_t)((int64_t)tmp)) << 32;
     env->macc[acc] = res;
     res = (uint32_t)env->macc[acc + 1];
     tmp = val & 0xffff0000;
-    res |= (int64_t)tmp << 16;
+    res |= ((uint64_t)((int64_t)tmp)) << 16;
     env->macc[acc + 1] = res;
 }
